update gcn attr for hiv bace

+++ models/graph_classifiers/GCN.py
@@ -83,7 +83,7 @@ class GNN_node(torch.nn.Module):
         
         ### computing input node embedding
 
-        h_list = [self.node_encoder(x)]
+        h_list = [self.node_encoder(x.float())]
         for layer in range(self.num_layer):
             h = self.convs[layer](h_list[layer], edge_index, edge_attr)
